Several fixes:
    Error handling in pod lifecycle
    Fix image pulling and add image pull progress tracking
    Update apptainer runtine directory management
    Fix MTU values and add forwards to host-bubble connectivity
    Fix certificates and bubble connectivity from the hosts
    Fix deletion of pod
    Add correct metrics, kube-proxy and work on dns error
    Fix certificates for multiple nodes
    Fix network setup inside container

+++ internal/compute/slurm/cancel.go
@@ -66,13 +66,18 @@ func CancelJob(args string) (string, error) {
 // KillProcessByPID terminates a process by its PID using kill command.
 // Returns an error if the process cannot be terminated.
 func KillProcessByPID(pid string) (string, error) {
+	pid = strings.TrimSpace(pid)
+	if pid == "" {
+		return "", ErrInvalidJob
+	}
+
 	/*
 	 Install trap for the signals INT and TERM to
 	 terminate the process and its children.
 	 Send SIGTERM using kill to the main process
 	 and wait for it to close gracefully.
 	*/
-	out, err := process.Execute("kill", "-9", pid)
+	out, err := process.Execute("pkill", "-P", pid)
 	if err != nil {
 		outStr := string(out)
